refactor: extract shared types into roux-core workspace crate (#13)
Create a Cargo workspace with roux-core crate containing all IPC model
types. src-tauri re-exports from roux-core. CLI and future tooling can
depend on roux-core directly.

Closes #13
