+++ Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+members = ["src-tauri", "crates/roux-core"]
+resolver = "2"
