+++ crates/roux-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "roux-core"
+version = "0.1.0"
+edition = "2021"
+description = "Shared types and models for Roux"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+specta = { version = "=2.0.0-rc.24", features = ["derive", "serde_json"] }
