+++ crates/roux-core/src/models/project.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
+pub struct Project {
+    pub id: String,
+    pub name: String,
+}
