+++ crates/roux-core/src/models/task.rs
@@ -1,0 +1,20 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
+#[serde(rename_all = "camelCase")]
+pub struct TaskDefinition {
+    pub id: String,
+    pub name: String,
+    pub description: String,
+    pub runner: String,
+    pub command: String,
+    pub keep_open: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
+#[serde(rename_all = "camelCase")]
+pub struct TaskGroup {
+    pub runner: String,
+    pub config_file: String,
+    pub tasks: Vec<TaskDefinition>,
+}
