+++ crates/roux-core/src/models/worktree.rs
@@ -1,0 +1,9 @@
+use serde::Serialize;
+
+#[derive(Debug, Clone, Serialize, specta::Type)]
+#[serde(rename_all = "camelCase")]
+pub struct Worktree {
+    pub path: String,
+    pub branch: String,
+    pub is_main: bool,
+}
