+++ src-tauri/Cargo.toml
@@ -43,6 +43,7 @@ reqwest = { version = "0.12", features = ["json"] }
 octocrab = "0.44"
 tauri-plugin-opener = "2.5.3"
 thiserror = "2"
+roux-core = { path = "../crates/roux-core" }
 specta = { version = "=2.0.0-rc.24", features = ["derive", "serde_json"] }
 specta-typescript = "0.0.11"
 tauri-specta = { version = "=2.0.0-rc.24", features = ["derive", "typescript"] }
