+++ src-tauri/src/tasks.rs
@@ -1,26 +1,7 @@
 use std::collections::HashMap;
 use std::path::Path;
 
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct TaskDefinition {
-    pub id: String,
-    pub name: String,
-    pub description: String,
-    pub runner: String,
-    pub command: String,
-    pub keep_open: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct TaskGroup {
-    pub runner: String,
-    pub config_file: String,
-    pub tasks: Vec<TaskDefinition>,
-}
+pub use roux_core::{TaskDefinition, TaskGroup};
 
 pub trait TaskDiscoverer {
     fn config_file(&self) -> &str;
